Compiled the swipe list with gradle(added dependencies and repositories) Added text to styles and drawable to help implement the swipes Added Swipe List view to the xml file in activity_list.xml added code to implement swipe list in ListActivity Everything runs fine in code but it does not seem to swipe on the phone - Made all the Swipe Code into Comments to deal with it later Added a new Relative layout to the listItem that contains the detailed Description of the events. Changed TestRound.xml to a rectangle to use it in the new RelativeLayout Created Textboxes in the new Relative Layout and assigned them in the layout Implemented the new textboxes and relative layout inside the Listview Adapter by adding them in the holder class and assigning them the value, still needs to be edited to change upon input. Event class still needs new variables,setters and getters to implement the new textboxes
-
Upon OnClick, the detailed description layout pops up and the summary description layout's contents are set to Invisible.
Event class is Implemented with its new variables.
Added the code to change the values of detailed description upon input in the Adapter class.(Made it into a comment for now though)
The Onclick to pop up detailed description is working at this point.
Implemented using clicks to: upon second tap,reverts to the original layout with the bubble summary description
Added comments to all codes